Move OV test to test_backends
