+++ tests/test_backends.py
@@ -1,14 +1,25 @@
 from __future__ import annotations
 
+import gc
 import json
 import os
 import tempfile
+from pathlib import Path
 
+import numpy as np
 import pytest
-from optimum.intel import OVModelForFeatureExtraction
-from optimum.onnxruntime import ORTModelForFeatureExtraction
+
+try:
+    from optimum.intel import OVModelForFeatureExtraction
+    from optimum.onnxruntime import ORTModelForFeatureExtraction
+except ImportError:
+    pytest.skip("OpenVINO and ONNX backends are not available", allow_module_level=True)
 
 from sentence_transformers import SentenceTransformer
+
+# openvino_unavailable = pytest.mark.skipif(find_spec("openvino") is None or find_spec("optimum.intel") is None, reason="OpenVINO must be installed for OpenVINO tests")
+# onnx_unavailable = pytest.mark.skipif(find_spec("onnxruntime") is None, reason="onnxruntime must be installed for ONNX tests")
+# backends_unavailable = pytest.mark.skipif()
 
 
 ## Testing exporting:
@@ -103,3 +114,46 @@ def test_openvino_provider() -> None:
 def test_incorrect_backend() -> None:
     with pytest.raises(ValueError):
         SentenceTransformer("sentence-transformers-testing/stsb-bert-tiny-safetensors", backend="incorrect_backend")
+
+
+# @pytest.mark.skipif(
+#     find_spec("openvino") is None or find_spec("optimum.intel") is None,
+#     reason="optimum-intel and openvino must be installed for OpenVINO test",
+# )
+def test_openvino_backend() -> None:
+    model_id = "sentence-transformers-testing/stsb-bert-tiny-safetensors"
+    # Test that OpenVINO output is close to PyTorch output
+    pytorch_model = SentenceTransformer(model_id)
+    openvino_model = SentenceTransformer(
+        model_id,
+        backend="openvino",
+        model_kwargs={"ov_config": {"INFERENCE_PRECISION_HINT": "f32"}},
+    )
+    pytorch_result = pytorch_model.encode(["Hello there!"])
+    openvino_result = openvino_model.encode(["Hello there!"])
+    assert np.allclose(openvino_result, pytorch_result, atol=0.000001), "OpenVINO and Pytorch outputs are not close"
+
+    with tempfile.TemporaryDirectory() as tmpdirname:
+        # Test that loading with ov_config file works as expected
+        config_file = str(Path(tmpdirname) / "ov_config.json")
+        with open(Path(config_file), "w") as f:
+            f.write('{"NUM_STREAMS" : "2"}')
+        openvino_model_with_config = SentenceTransformer(
+            model_id,
+            backend="openvino",
+            model_kwargs={"ov_config": config_file},
+        )
+        # The transformers model is an Optimum model with an OpenVINO inference request property
+        assert openvino_model_with_config[0].auto_model.request.get_property("NUM_STREAMS") == 2
+
+        # Test that saving and loading local OpenVINO models works as expected
+        openvino_model_with_config.save(tmpdirname)
+        local_openvino_model = SentenceTransformer(
+            tmpdirname, backend="openvino", model_kwargs={"ov_config": {"INFERENCE_PRECISION_HINT": "f32"}}
+        )
+        local_openvino_result = local_openvino_model.encode(["Hello there!"])
+        assert np.allclose(
+            local_openvino_result, openvino_result
+        ), "OpenVINO saved model output differs from in-memory converted model"
+        del local_openvino_model
+        gc.collect()
